Implementing edge swap and rearranging files
