+++ go.mod
@@ -1,0 +1,3 @@
+module IMO
+
+go 1.24.1
