--- zad1/go.mod
@@ -1,3 +1,0 @@
-module zad1
-
-go 1.24.1
