+++ zad2/README.md
@@ -1,0 +1,1 @@
+# IMO
